Rename file with camel case
