+++ scriptToSaveVideo.py
@@ -3,9 +3,9 @@ import pafy
 
 from Common import url as Url
 
-PATH = "video/cambridge9.avi"
+PATH = "video/Derry/derry1.avi"
 
-video_pafy = pafy.new(Url.CAMBRIDGE['Path'])
+video_pafy = pafy.new(Url.DERRY['Path'])
 play = video_pafy.getbest()
 cap = cv.VideoCapture(play.url)
 
@@ -16,13 +16,16 @@ size_frame = (frame_width, frame_height)
 
 video = cv.VideoWriter(PATH, cv.VideoWriter_fourcc(*'MJPG'), 10, size_frame)
 
+window_name = "Save video in progres..."
+cv.namedWindow(window_name)
+
 while cap.isOpened():
     ret, frame = cap.read()
 
     if ret:
         video.write(frame)
         frame = cv.resize(frame, (650, 650))
-        cv.imshow("OpenCv", frame)
+        cv.imshow(window_name, frame)
         cv.waitKey(1)
 
         if 0xFF == ord('q'):
